+++ NEWS.md
@@ -1,5 +1,6 @@
 simhelpers 0.3.0.9999
 =======================
+* Added support for bias-corrected and bias-corrected-and-accelerated (BCa) bootstrap confidence intervals.
 * Corrected an error in the documentation of `bundle_sim()`.
 
 simhelpers 0.3.0